% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.R
\name{add_lines}
\alias{add_lines}
\title{Add Line Features (Classes) to Landscape}
\usage{
add_lines(landscape, f, q, overlap)
}
\arguments{
\item{landscape}{\code{matrix} a landscape matrix}

\item{f}{\code{vector} (length based on number of desired line classes) frequency of given class, if \code{q[i] = FALSE} the frequency \code{f[i]} indicates the rate of patches connected by line features (ordered by area descending)}

\item{q}{\code{vector} (length based on number of desired line classes) used for binning classification if \code{q[i] = FALSE} given class created by straight lines}

\item{overlap}{\code{vector} landscape class to be overlapped with the given line class, if \code{q[i] = FALSE} patches of given class (\code{overlap[i]}) will be connected by the created line features}
}
\value{
landscape with added line features
}
\description{
description ...
}
\examples{
add_lines(matrix(0, 100, 100), c(.005), c(.05), c(0))
}
