% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate Landscape Parameters}
\usage{
estimate_parameters(landscape, curve = NA, line = NA)
}
\arguments{
\item{landscape}{\code{matrix} with integer values representing different land-cover classes}

\item{curve}{\code{list} each element of the list contains a) class label and b) overlapping (polygon) classes}

\item{line}{\code{list} each element of the list contains a) class label and b) connected (polygon) classes}
}
\value{
estimation of parameters of the given landscape
}
\description{
Used to get parameters for virtual landscape creation functions \code{create_landscape} and \code{add_lines}
}
\examples{
estimate_parameters(landscape, 5, 4)
}
