% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimator}
\alias{estimator}
\title{Landscape parameter estimator}
\format{
An object of class \code{list} of length 2.
}
\usage{
estimator
}
\description{
Model for parameter estimation
}
\keyword{datasets}
