% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape}
\alias{landscape}
\title{Sample Landscape}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 100 columns.
}
\usage{
landscape
}
\description{
Sample random landscape created using \code{{landsimR}} package functions:
\code{create_landscape} & \code{add_lines}
}
\keyword{datasets}
