% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-class.R
\name{LandscapeObj}
\alias{LandscapeObj}
\title{construct a \code{landsimR} landscape \code{LandscapeObj} type}
\usage{
LandscapeObj(landscape, classType)
}
\arguments{
\item{landscape}{\code{matrix} landscape matrix}

\item{classType}{\code{numeric} class types: polygon = 1, curve = 2, line = 3}
}
\value{
\code{Landscape} landscape object
}
\description{
A short description...
}
\examples{
LandscapeObj(landscape, c(1, 1, 1, 1, 2, 3))
}
