% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-class.R
\name{plot.LandscapeObj}
\alias{plot.LandscapeObj}
\title{plot \code{landsimR} landscape \code{LandscapeObj}}
\usage{
\method{plot}{LandscapeObj}(x, ...)
}
\arguments{
\item{x}{\code{LandscapeObj}}

\item{...}{additional arguments of \code{image} or \code{plot}}
}
\description{
A short description...
}
\examples{
plot(landscape, axes = FALSE)
}
