% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate Landscape Parameters}
\usage{
estimate_parameters(landscape)
}
\arguments{
\item{landscape}{\code{matrix} with integer values representing different land-cover classes}
}
\value{
estimation of parameters of the given landscape
}
\description{
Used to get parameters for virtual landscape creation functions \code{create_landscape} and \code{add_lines}
}
\examples{
estimate_parameters(matrix(sample(c(1,2,3), 10000), 100, 100))
}
