% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsimR-package.R
\docType{package}
\name{landsimR-package}
\alias{landsimR}
\alias{landsimR-package}
\title{landsimR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-kucera/landsimR}
  \item \url{https://m-kucera.github.io/landsimR/}
  \item Report bugs at \url{https://github.com/m-kucera/landsimR/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
